+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - fixed issue with boolean require validation
+- get correct option labels on store view 
 ### Added
 - select options validation: checking if option id exists before its saved